feat: 팔로우 이벤트 처리 구현

+++ consumer/src/main/java/net/happykoo/kafkanoti/task/LikeRemoveTask.java
@@ -1,7 +1,9 @@
 package net.happykoo.kafkanoti.task;
 
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import net.happykoo.kafkanoti.domain.notification.LikeNotification;
+import net.happykoo.kafkanoti.domain.notification.Notification;
 import net.happykoo.kafkanoti.domain.notification.NotificationType;
 import net.happykoo.kafkanoti.event.LikeEvent;
 import net.happykoo.kafkanoti.service.NotificationService;
@@ -14,16 +16,19 @@ public class LikeRemoveTask {
   private final NotificationService notificationService;
 
   public void processEvent(LikeEvent event) {
-    LikeNotification notification = (LikeNotification) notificationService.findByTypeAndPostId(
+    Optional<Notification> notificationOpt = notificationService.findByTypeAndPostId(
         NotificationType.LIKE,
-        event.getPostId()).orElseThrow(IllegalArgumentException::new);
+        event.getPostId());
 
-    notification.removeLiker(event.getUserId());
+    if (notificationOpt.isPresent()) {
+      LikeNotification notification = (LikeNotification) notificationOpt.get();
+      notification.removeLiker(event.getUserId());
 
-    if (notification.getLikerIds().isEmpty()) {
-      notificationService.deleteById(notification.getId());
-    } else {
-      notificationService.save(notification);
+      if (notification.getLikerIds().isEmpty()) {
+        notificationService.deleteById(notification.getId());
+      } else {
+        notificationService.save(notification);
+      }
     }
   }
 }
